+++ src/OpenSage.Mods.Generals/Gui/GeneralsControlBar.cs
@@ -321,7 +321,7 @@ namespace OpenSage.Mods.Generals.Gui
                                         //TODO: figure this out correctly
                                         var selection = context.Game.Scene3D.LocalPlayer.SelectedUnits;
                                         var objId = context.Game.Scene3D.GameObjects.GetObjectId(selection.First());
-                                        order.AddIntegerArgument(objId);
+                                        order.AddIntegerArgument((int)objId);
                                         var upgrade = commandButton.Upgrade.Value;
                                         order.AddIntegerArgument(upgrade.InternalId);
                                     }
